Fix error when ortho dirs are completely or partially missing

+++ autoortho/downloader.py
@@ -243,12 +243,12 @@ class OrthoRegion(object):
             log.info(f"Region {self.region_id} version {self.latest_version} not downloaded!")
             return
 
-
         if self.ortho_dirs:
             log.info(f"Detected existing scenery dirs for {self.region_id}.  Cleanup first")
             self.cur_activity['status'] = f"Detected existing scenery dirs for {self.region_id}.  Cleanup first."
             for o in self.ortho_dirs:
-                shutil.rmtree(o)
+                if os.path.exists(o):
+                    shutil.rmtree(o)
 
         log.info(f"Ready to extract archives for {self.region_id} v{self.latest_version}!")
         self.cur_activity['status'] = f"Extracting archives for {self.region_id} v{self.latest_version}"
